[UI 제작] 플레이어 경험치 UI바 꽉 찼을 때 초기화 작업

+++ Source/RiseOfSun/Public/RSPlayer.h
@@ -25,7 +25,7 @@ public:
 
     // ---------- 여기부터 EXP 관련 추가 ----------
     UFUNCTION(BlueprintCallable)
-    void AddEXP(int32 ExpAmount);
+    void AddEXP(float ExpAmount);
 
     UPROPERTY(BlueprintAssignable)
     FOnEXPChanged OnEXPChanged;
